+++ ConvertZZ/Moudle/StreamString.cs
@@ -32,8 +32,8 @@ namespace ConvertZZ.Moudle {
         public int WriteString(string outString) {
             byte[] outBuffer = streamEncoding.GetBytes(outString);
             int len = outBuffer.Length;
-            if (len > UInt16.MaxValue) {
-                len = (int)UInt16.MaxValue;
+            if (len > ushort.MaxValue) {
+                len = ushort.MaxValue;
             }
             ioStream.WriteByte((byte)(len / 256));
             ioStream.WriteByte((byte)(len & 255));
@@ -44,8 +44,8 @@ namespace ConvertZZ.Moudle {
         public async Task<int> WriteStringAsync(string outString) {
             byte[] outBuffer = streamEncoding.GetBytes(outString);
             int len = outBuffer.Length;
-            if (len > UInt16.MaxValue) {
-                len = (int)UInt16.MaxValue;
+            if (len > ushort.MaxValue) {
+                len = ushort.MaxValue;
             }
             await ioStream.WriteAsync(new byte[] { (byte)(len / 256) }, 0, 1);
             await ioStream.WriteAsync(new byte[] { (byte)(len & 255) }, 0, 1);
